Better api and comments

+++ todoList/server/index.js
@@ -29,7 +29,7 @@ class Database {
             id: nextId++,
             title: todo.title,
             isDone: false
-        }
+        };
     
         this.todos.push(newTodo);
     
@@ -55,7 +55,7 @@ app.use(express.json());
  * Returns all todos
  */
 app.get('/todos', (req, res) => {
-    res.json(database.findAll());
+    res.json(ok(database.findAll()));
 });
 
 /**
@@ -63,14 +63,17 @@ app.get('/todos', (req, res) => {
  */
 app.get('/todos/:todoId', (req, res) => {
     const id = parseInt(req.params.todoId, 10);
+    if (isNaN(id)) {
+        return res.json(error(`Invalid id '${id}'`));
+    }
 
     const todo = database.findById(id);
 
     if (!todo) {
-        return res.status(404).send("Sorry can't find that!")
+        return res.json(error(`Unable to find todo with id '${id}'`));
     }
  
-    res.json(todo);
+    res.json(ok(todo));
 });
 
 /**
@@ -78,19 +81,22 @@ app.get('/todos/:todoId', (req, res) => {
  */
 app.put('/todos/:todoId', (req, res) => {
     const id = parseInt(req.params.todoId, 10);
+    if (isNaN(id)) {
+        return res.json(error(`Invalid id '${id}'`));
+    }
 
     const newData = {};
 
     if (req.body.title) {
         if (typeof req.body.title !== "string") {
-            return res.status(400).send("'title' should be of type string");
+            return res.json(error("'title' should be of type string"));
         }
         newData.title = req.body.title;
     }
 
     if (req.body.isDone) {
         if (typeof req.body.isDone !== "boolean") {
-            return res.status(400).send("'isDone' should be of type Boolean");
+            return res.json(error("'isDone' should be of type Boolean"));
         }
         newData.isDone = req.body.isDone;
     }
@@ -98,12 +104,12 @@ app.put('/todos/:todoId', (req, res) => {
     const todo = database.findById(id);
 
     if (!todo) {
-        return res.status(404).send("Sorry can't find that!")
+        return res.json(error(`Unable to find todo with id '${id}'`));
     }
 
     database.update(todo, newData);
  
-    res.json(todo);
+    res.json(ok(todo));
 });
 
 /**
@@ -111,16 +117,19 @@ app.put('/todos/:todoId', (req, res) => {
  */
 app.delete('/todos/:todoId', (req, res) => {
     const id = parseInt(req.params.todoId, 10);
+    if (isNaN(id)) {
+        return res.json(error(`Invalid id '${id}'`));
+    }
 
     const todo = database.findById(id);
 
     if (!todo) {
-        return res.status(404).send("Sorry can't find that!");
+        return res.json(error(`Unable to find todo with id '${id}'`));
     }
  
     database.remove(todo);
 
-    res.json("OK");
+    res.json(ok());
 });
 
 /**
@@ -131,25 +140,39 @@ app.post('/todos/add', (req, res) => {
 
     if (req.body.title) {
         if (typeof req.body.title !== "string") {
-            return res.status(400).send("'title' should be of type string");
+            return res.json(error("'title' should be of type string"));
         }
         newData.title = req.body.title;
     } else {
-        return res.status(400).send("'title' can not be empty");
+        return res.json(error("'title' can not be empty"));
     }
 
     if (req.body.isDone) {
         if (typeof req.body.isDone !== "boolean") {
-            return res.status(400).send("'isDone' should be of type Boolean");
+            return res.json(error("'isDone' should be of type Boolean"));
         }
         newData.isDone = req.body.isDone;
     }
 
     const newTodo = database.add(newData);
 
-    res.json(newTodo);
+    res.json(ok(newTodo));
 });
+
+function ok(data) {
+    return {
+        status: "OK",
+        data: data
+    }
+}
+
+function error(message) {
+    return {
+        status: "ERROR",
+        message: message
+    }
+}
 
 app.use(express.static('./'));
 
-app.listen(port, () => console.log(`Example app listening on port ${port}!`));
+app.listen(port, () => console.log(`Example app listening on port ${port}!, try http://localhost:3000/`));
